Cleanup/unify actions in Linux & Windows examples.
The goal of the actions was to unify the Linux and Windows examples. There should be no difference between windows/linux versions of isisuc.tab, isisac.tab, isisuc_utf8.tab, isisac_utf8.tab. During this compare a lot of orphans were found and deleted.
Note that many  binary files are different and not touched by this action