Deleted duplicate & semi duplicate isisuctab_utf8.tab
The default location of this file is in the "bases" folder. Individual databases may have a specific version. This coded in utilities/vmx_fullinv.php.
All files above were exact duplicates or duplicates with a typo and different spacing.
Maintenance will be easier if this kind of duplicates is removed