Deleted duplicate & semi duplicate isisactab_utf8.tab
The default location of this file is in the "bases" folder. Individual databases may have a specific version. This is coded in utilities/vmx_fullinv.php.
The file in the "bases" folder contained a superset of the information found in the local folders. Inspection of the differences gave the very strong impression that these differences were not functional but due to historic growth on different platforms.
Maintenance will be easier if this kind of duplicates is removed