error when searching
- buscar.php: updated by guilda
- fmt.php: updated by guilda
- ifpro.php and inicio_main.php : not updated. Content was already in git
Comment Guilda: The problem with search expressions is the following: in HTML double quotes or the apostrophe are used to delimit variables, but both characters are necessary since expressions must be delimited between "..." to allow the and, or and not  to be part of the expression and the ' is used in names and other elements in different languages. I did some changes that can help to solve this issue. Please test.

In version 2.2, the option to enclose each term in the search expression between "..." was added to incorporate the use of and or and not as part of the term. This can cause some problems in some search processes that have not yet been corrected. Please advise.