Improve show of isofile
A new file "mx_show_iso.php" disentangles the function to show the content of an iso file from mx_dbread.php (which shows also mst files).  This enabled implementation of display UTF/ISO and solution of the problem that mx does not hnour the "from" clause for iso files.
- exporta_txt_ex.php : use mx_show_iso.php + moved action columns after filename
- up2.gif: new icon to indicate "back_to_top"
- menu_bar.php: Improved text for Import ISO+ disabled mx_dbread.php
- mx_show_iso.php: new function
- vmx_import.php:  use mx_show_iso.php+ solved bug for files without extension