Correct import and export for isotag1
Importing with option isotag1=3000 for non-marc iso files resulted in invisible additional fields in the database.  This commit is solves this for non-marc databases. For marc databases the export is enhanced but the import needs additional work.
In all files several fixed strings are replaced by translatable strings.
Detailed modifications:
- exporta_txt.php:  always preselect mx+improved strings
- exporta_txt_ex.php:Improved command for marc database and non-marc database
- clean_db.php: removed option isotag1=3000
- inc_check_line-end,php: no error if file has 1 line without line-end. This is produced by export for marc.
- vmx_import.php: removed option isotag1=3000