Remove password from $_SESSION. Second reduction of usage of password
The password was set into the session object and also retrieved from it in various scripts. These scripts did not use the password but echoed it sometimes  in an URL for a next action (which did not use the password too).  This is a security leak.
Major modification of the scripts in this commit remove code to set/retrieve the password from the session object.
Minor modifications are listed in the modification comments of the scripts.
The deleted scripts were not used.