Repair non-working option in conf-abcd.php
The option was configuration of opac. The addressed file was central/dbadmin/opac_configure.php. The code in this file was not working due to wrong paths AND it did not comply with the instractions for configuration of OPAC on the wiki (http://abcdwiki.net/EN_OPAC-ABCD_Tutorial_de_configuraci%C3%B3n).
Actions:
- opac_configure.php : deleted
- conf_abcd.php: Removed reference to opac_configure + Cleanup html, cleanup code and lineends