Update global change
This is a first step to make global change more user friendly,reliable and predictable. The unused existing message translations indicate that more is probably necessary.
This commit does NOT change anything to the wxis scripts that do the real work.
Note that the cisis windows code reacts different from the linux code. See issue  #126
Change details are in the Modifcations header  of the code. Highlights:
- c_global.php: No longer works with a parallel active pop-up but follow up screen. Many errors are already detected.
- c_global_ex.php: standardized code and generated html
- leer_all.php: Show always arrows as text start/terminator in stead of  strange characters
- soporte.tab: Added some messages