Add utility to match the entries of an ISO file with the FDT
The records of an ISO export file may contain entries/fiields with tags that are not in the FDT of the databases. The import process will load them without comments into the database. In order to prevent that these tags are loaded into the database this utility can be used to create an ISO file where all fields with tag values that are not present in the FDT are no longer present.