Update subfieldcodes for FDT option OD in dibujarhojaentrada.php
OD is about a date and an operator. If no codes are set in the FDT a default is applied. This default was wrong and is corrected by this commit.
The field will now show:
"^d20210720 07:18:05^oabcd" : The ^d for the date and ^o for the operator.
Note 1: This was also the behavior for ABCD1.4 and earlier
Note 2: The date is (and was always) the first field , followed by the operator