New browse.php
New Browser!

Dear friends, I present a new proposal for the browser.php present in several parts of ABCD.

I made a lot of graphic changes, don't be scared, but I would like to replace little by little the old icons and organize them in one place, but this is for the future. Below I describe the main elements of the new browser.php

- "Advanced Search" button: its purpose is to replace the select that used to trigger the index. When I clicked on the "Index" button without previously selecting a field, a warning appeared. It seemed annoying to me.
- Search form: was improved to accept combinations of terms received by the advanced search.
- "Reset" button: clears all filters and search expressions.
- Select "show": allows you to choose how many records are shown in the table.
- Select "order by": allows you to sort the table by a column, provided that the sort.tab file exists (already included in all bases).
- Select "reverse": allows you to sort the selected column in ascending or descending order.


Table:
- The table has undergone some visual changes. With the intention of bringing more lightness to the ABCD layout, I removed the colors, shadows and borders.
- I replaced the gif or png icons with flat icons using Fontsawesome.


- The page navigation by numbers seems more comfortable to me as it allows jumping between pages without having to go one by one or go straight to the end. The current script is in PHP and no longer in JavaScript.

- Show deleted" button: to replace the checkbox that used to display deleted records in wxis "mfnrange" mode. Deleted records are only displayed by "mfnrange", i.e. there is no other way to display them, after all they are not indexed. Therefore they are not searchable. In a very large database it would be very cumbersome to locate these records, so pressing this button toggles the display to deleted only and hides the active records.

- "Preview" button: this is actually "print", but I didn't want to make an entry in the "lang" files at this point. This button triggers the printing of what is showing on the screen. To work, it needs the file "tbBASE_html.pft" to exist in the base.

- "Spreadsheet" button: this is the classic button that exports the view into an Excel file.


As mentioned above, this new layout with new functionality impacted several points in the program.

I created the file "dataentry/print.php" - a copy of imprimir_g.php, but with print-friendly elements.

I created at the root the directory "assets" to install FontAwesome, but it will serve in the future to organize the image, css and javascript files that are scattered in the directories.

The bases below needed adjustments, standardization and even the creation of their formats in some languages:
- acces
- copies
- providers
- purchaseorder
- reserve
- suggestions
- suspml
- trans
- users

Also the CSS files were changed, so a CTRL+R is needed.

Sorry for the more than 400 files changed at once, but I think the result is satisfactory, although it needs some adjustments.

Best regards.