new look - p1
Dear Friends,
I am working hard on reorganising the ABCD layout. With that, I'm taking the opportunity to implement some graphical changes.

With the inclusion of Font Awesome in the Browser it was easier to create buttons and replace some icons that look like Windows Vista (nothing against it, but its time has passed).

The work is far from being concluded. There are many css, images and JS files scattered around the program.

My intention is to unify everything in the /assets directory. 

Since I moved a lot of files, I decided to do a commit before continuing.

It was moved 415 files, but most of them only had changes in the images path.

Some files in the /acquisitions folder were corrected.

The /central/images folder was moved to /assets/images, as well as the css files.

I made a big modification in /dbadmin/menu_mantenimiento.php, I included an iframe to load the navbar scripts. The iframe automatically adjusts itself to the loaded file.

In the next updates I intend to move the folders:
/central/imgs
/central/dataentry/img
/central/dataentry/js

Best regards!