Improve reaction for invalid language
The current version detects the default langauge from the browser. This works  fine as long as the language exists in ABCD.
This commit checks if the browser language is present and switches to the ABCD default language if applicable.
The messages are now also visible in the browser so the user is informed. (previous status was that the message was only visible in the page source and not directly on screen)