Adjustment for the select to point to the default language.
I remember that in some version of ABCD there was a function that set English as default in the absence of the directory of some language, so I added the option without realizing that this function did not exist.

About the adopted solution, I don't know if it is safe to put the server path in the index. In demo installations it will show the server path publicly.