Update document import
This commit is the first reaction on the recent comments. Not all comments are properly addressed here, more will follow.
Details:
- If no tika is found a message shows a download indication
- Multiple tika jars can be located in cgi and are shown in a dropdown menu. Implies that existing and new tika's can be used at will
- A "test" link opens a pop-up that informs about the tika versions
- The split size target can now be controlled by the menu (default set to 80% of record size.
- The filename can be truncated, even to 0 characters
- The sanitized filename is free from spaces+ windows filename restricted characters+ rfc 3986 gen-delims/sub-delims
- Some small other enhancements.
This version is tested in windows  and linux. Both work OK with "standard" filenames. With UTF-8/Windows 1252 Windows needs some enhancements and Linux needs more enhancements. (Most serious on linux is the complaint of tika about a malformed URL which indicates that something is wrong with the filename)
The database scheme is not modified (no need to adapt fdt/fst/fmt/pft)