Adjustments to the "Suggestions" database and more
# Suggestions database
In this first evaluation, some fields of the suggestions database were corrected. This database is responsible for storing the acquisition requests for the collection.

This module is composed by 3 databases and all of them will be reviewed in the next commits: 
- providers
- purchaseorder
- suggestions

These databases should exchange information with each other, but so far the fields are not perfectly interlinked.

# Input type number
To improve the filling of numeric fields, avoiding errors or typing of another type of data the field type "number'" was inserted in the FDT