Update to improve dubcore functionality
This commit is triggered by the given fact that dubcore requires inverted file generation by /m and the fact that DB picklists give no results when this option is used.
This implies that the quick search (A-Z button) does not function and that record modification using a DB picklist does not work too.

The key of the solution is a variant of ifp.xis: ifp_slashm.xis. This is a new script, only applied in modified and tested scripts. Other scripts using ifp.xis still use the original.
The new script performs like the original for databases indexed without /m (the default). The new script does not completely solve the problem: There are limitations to the indexing method used in the FST: Indexing method 5 By subfield with prefix works acceptable. Other indexing methods do not work or work incorrect.

During testing several other bugs were detected and addressed. The code is reshuffled to allow debug output on screen in stead of hidden in the webpage.

Other major changes:
 - capturaclaves.php: background white (class formContent)+ div-helper+correct text in breadcrumb+ A-Z buttons left of selection window (like in quick search)+navigation buttons as in diccionario and quick search. The Preview and Print buttons are not touched (and may need addional work). Functionality of this script with regard to barcode and thesaurus is left in place, but not tested.
- alfa.php: Improved switching of databases in the copy function + message if no principal entry is found in the FDT
- captuarclaves
Minor modifications:
 - fullinv.php: If the script applies the /m modifier a blue message is shown to warn the user for DB picklists
- diccionario.php: use also ifp_slashm and better "enter" button