Transition review for the new look
Following the work of a few months ago, the new CSS is being organised to serve developers and database managers. In this commit are being uploaded the colors.css and buttons.css files that correspond to:
- system colour palette
- general colour palette - for use on demand
- text colours
- background colours
- colored buttons
- different button sizes

This update happens in the middle of the Acquisition module remodeling due to the need of the presence of some elements to give continuity to this work.