Update on the configuration screen
This commit is due to the previous commit where the varied colour palette was made available for general use.

The structural colours of the system can be changed through the interface by the end user.

To allow dynamic customisation for the end user, the file: http://localhost:9090/central/dbadmin/editar_abcd_def.php has been updated

In the file colors.css the classes have been eliminated (to avoid confusion):
  --blue-abcd: #003366; /* HEADING
  --bgmoduleInfo: #336699;
  --bgsectionInfo: var(--bgmoduleInfo); /* SECTIONINFO
  --cat-color: #CCCCFF;
  --acq-color: #96EAC8; --loan-color: #99
  --loan-color: #99FFCC;