Preset configured language in login language selection dropdown
Note that the login screen is by default shown in the language of the browser.
The language dropdown took that also as preset, ignoring the preset language in abcd.def.
This modification preselects the preset language from abcd.def in the language dropdown.

The initial screen takes the browser language and if that is not available the fallback. This is the language set by config.php (may be the value set in abcd.def). Missing translations are given by the built-in 00 (if central/lang/00 is up to date)

The