Favicon update
The white favicon was not working well in browsers with clear top bars.

A .svg file has been included as the main favicon, because it can be edited in notepad and the colour changed by modifying the "fill" selector

It included coloured icons to be displayed in shortcuts and favourites in some browsers or devices.

Extra: in Firefox the favicon has a slight animation.