New revision in the ABCD.DEF editor
In this revision the problems caused by missing colours in the layout have been corrected.

Even if the user has a file abcd.def in his base folder without colour parameters the script css_settings.php will be able to put the correct colours in the layout.

2 fields were created to upload logos: institution logo at the top of the screen and the logo of the institution responsible, at the right at the bottom.

Because of the possibility of uploading custom logos, the upper text in <h1> next to the logo has been removed because it was occupying and causing the layout to break. The text was placed inside the image as a title attribute. It is displayed when the mouse passes over it.

The title of the institution repeats in the footer in the textual version. In its absence, the name of the ABCD in several languages keeps appearing every time you update it (a bit of fun here)

Those parameters called "LEGEND" were eliminated, because their names were not clear.

This is another step towards end-user autonomy, where it is possible to adjust the layout according to the visual identity of the institution without the need for webdesign skills.

The suggestion of including a button to recover the default colours is still being evaluated.