More security
Enabling the use of the .htaccess file in the root of the public folder in Apache configuration files.

Changed from:
AllowOverride None

To:
AllowOverride All

Addition of the default .htaccess, where its behaviour is:
If any user accesses any directory that does not have an index file, naturally the browser will display the main index.php

The file also blocks access to the config.php file and htacces itself.