The new settings editor
The new settings editor now displays the directories.

In the "folders" option the user must choose a directory to store the logo files - the result of this action will be extended to iah and opac soon.

After choosing a directory, the Logos tab will appear.

An aesthetic resource was included in the display of abcd.def to give time to save and return to the form with the new rules. The reason for this is the configuration of the upload directory that enables the logos tab.

Some more adjustments were made in the files related to this action.