New look 5
This commit introduced the new back/cancel/save buttons in remaining central/dataentry scripts. If applicable the inc_div-helper is also applied.

Script toolbar_record no longer shows empty toolbars. (the strange blue block in the right upper corner in some screens)

An attempt is done to normalize the generated html. As these scripts are very large and complicated this will not be the last time that it is needed.