Improve highlight after search
The prologact.pft's referenced iah/js/highlight.js. This is corrected to central/dataentry/js/highlight.js This ensures independence from iAH. Note that this must be mentioned in the upgrade procedure.

Both scripts showed a java error for a sometimes missing reference. This is corrected in the central version.

highlight.js has some translations for characters with diacritics. Some translation were wrong and some were missing. Both are corrected in the central version