Move default barcode font file to assets
File BAR39EH.TTF was located in the script folders.
It is moved to a new folderin assets: assets/barcodefonts.
This folder is intended to be traversed to show font-files for the current ABCD installation (we can add TTF or zip files with fonts or the local system manager can do this to please his users)