Adjustments in MARC database indexing for Opac
The prefixes "CR_" and "SU_" were included to be similar to the database dublincore.

IMPORTANT: For the advanced search to work properly, the prefixes of the fields need to be the same in all databases.

Note: until this commit the prefix "CR_"doesn't work in the advanced search, strangely there is something eliminating this letter combination.