Update .htaccess
Included the parameter "php_value max_input_vars" to change the parameter "max_input_vars" of php.ini which limits the sending of forms with a maximum of 1000 fields.

Due to the size of MARC, MARCUNI and RDA databases, this parameter is necessary to avoid errors when saving a record in cataloguing.