Preparation for fundamental improvement of barcode scripts
This commit adds new references in menu "Update database definitions": to create a label & barcode table (to replace the fixed built in 3 types) and to configure a selected type from that table.
Currently empty scripts: next commits.

The layout of the  "Update database definitions" menu is beautified and an invalid permission check is removed