Change to the Opac structure
This commit modifies the way Opac works at the public level (htdocs).

Starting with this commit, configuration files related to some database will be inside the database in a folder called opac.

In this model the MARC and DUBCORE databases are being used.

In this group of files a new variable was added that will allow the change of the source of the .par files