Users Base for Linux - updated!
Updates have been made to the test user data.
- 2 users are with active registration, that is, without expired registration;
- 1 user has expired registration, but can still make loans.

The photos of people who do not exist generated by the site were changed: https://this-person-does-not-exist.com/en

The display of photos was adjusted to display by default the images without the dr_path.def file, but the upload needs to be adjusted to a standard operation.