Number of copies to be added blocked
The "Number of copies to be added" field is used to automatically create the numbers in the "Inventory number" field, but only numeric values are allowed.
In case of manual entry of several "Inventory number" (barcodes), the "Number of copies to be added" field will be ignored, even if it has been filled in.