First version of the new MySite
The MySite module is being built with the focus on mobile phones, as it is an interface that serves users of a library, for example.

First version of the new MySite

The MySite module is being built with the focus on mobile phones, as it is an interface that serves users of a library, for example.

You can test it with the user: rosa and the password: rosa