Write empty shortcut.pft files
For iAH there needs to be a file 'shortcut.pft', containing links - if any, can also be left empty - to external PFTs e.g. for production of bibliographic lists like for Reference Manager, which needs to be correctly referenced to in the DBN.def file.
iah_save.php:
  - Gets the value of FILE SHORTCUT.IAH,
  - replaces %path_database%,
  - If the value does not contain %lang% an empty file is created if not present
  - if the value contains %lang% then for each iah languages an empty file is created if not present
iah_edit.php:
  - Some small error corrections. Not ideal but it works now