Commit to remove English language files
The English language is delivered by code stored in central/lang/00.
Goal is that the English language is always present, up to date and in-sync with the script code
Additional English language files will compromise this goal and should not be present in the default situation.

Note that local administrators still can add local differences if they create their own language files. Advise is that these files contain only entries for local differences to keep maintenance to a minimum.