Improve iso import/export for marc databases
The iso export code for marc databases no longer uses the mx option iso=marc=filename. Only outisotag1 is used. Effect is that correct sio files with record length=80 are produced

The iso import code has an extra option to specify the application of isotag1 in the import command.
When selected the import command uses isotag1=3000 and import the leader info of marc databases.
Note 1: The program does NOT help the importer with any detection if leader information is present or not