Repair vanishing record values when word script was used
This commit repairs the unintended clearance of record values when the word "script" appeared in the value of such a record.  Dependent of the case  only the attribute value was cleared and in case of validation errors much more value were cleared. The root cause was a test on the word "script" in the attribute value, probably added in an attempt to avoid injection of javascript in an attribute value. The test is  removed completely: this and similar too simple tests cause more harm then protection.
Detailed modifications:
  * actualizarregistro.php:
    * Use div-helper, improve human readability
    * removed trailing \n before explosion of value
    * Added function parameter $debugparameter as help for future debugging
  * fmt.php:
    * Don't set arrHttp values to **INVALID** if they contain the string "script"
    * debug parameter for ActualizarRegistro
    * check existence of indexvalue in creation of wks values in case "reintentar"
  * fmt_begin_end.php:  debug parameter for ActualizarRegistro
  * recval_check.php: improve html syntax