Repair layout problems in cataloging
Solves the problem that it was sometimes not possible to scroll down to the bottom while stepping the records (missing footer or more) or editing the record (missing buttons or more).
Solved by better computation of the height of iframe main.
The other modifications in inicio_main.php are moving valid css attributes to styles.css and removing invalid attributes.
Note that the layout is very slightly improved and the menu bar may show a scroll bar too in case of  a window that is much  too small.

Solves the problem that in certain circumstances in an edit session the dropdown menus shifted to a step to the right, so the layout of the form looked clumsy.
Modified by settings of table-fdt-three in styles.css