Commit to show back button after successful update of administrative record
Saving the edit of a record for administrative databases like Operator, Borrower, Transaction, Suggestion, Provider+ several more, shows now the updated record with in the heading a Back button. This button was missing. This button returns to the Home page of Cataloging/Circulation/Acquisition as appropriate.
Additionally following improvements are done too:
  * The result screen shows now also the institutional info and a better located helper.
  * The text in the breadcrumb for the edit/create and result are improved
  * Some code cleanup