Improve FDT update and cleanup dhtmlX related coding
The reason for this change was adding a message in FDT update to indicate that the OD entry requires a non-zero number in the "cols" column to indicate the number of stored updates.
It appeared that the grid was not properly exposed. Further investigation showed many (sometimes hidden) errors in scripts using the functionality in folder dhtml_grid. Solution was: always use the css from /assets + move images also to /assets + sanitize several php scripts to use this functionality properly.  Visible effects are: Table sort and moving rows works now, new look and translations are extended in these scripts.