Improve gizmo support
The utilities menu "Create gizmo" is replaced by "Manage gizmo": Manage the gizmo databases and related files in the data folder of the selected database. The main script shows options to upload a file, Create a gizmo and the list of gizmo related files (.mst,.iso,.seq) with options to view/edit/remove these files. Exporting gizmo databases is added, to iso and a text file (.seq). Notable script details:
- menubar.php: Replaces Create gizmo by Manage gizmo
- manage_gizmo.php: new. The main gizmo management script
- create_gizmo.php: More interactive options, more robust
- export_gizmo.php: new. Exports to iso or text (using mx option fix)
- edittxtfile.php: new. Interactive edit of a text file with preservation of the original encoding, line endings and possible html codes.
- fullinv.php: Add options to add gizmo's for the inversion. Better readable output . Check lineendings of isisuc.dat
- vmx_import_iso.php: Add option to add gizmo's for the import process
- upload_wrkfile.php: Caller can specifiy other upload location (was fixed to /wrk)