Update editar_abcd_def.php
Creates the LOGIN_ERROR_PAGE parameter to allow hosts where the login page is allocated to another directory to return to this directory in the event of an error. Today, when the OPAC is the homepage, when the login fails the user is sent to the Opac, instead of returning for another login attempt.